Time: 0 ms (100.00%) | Memory: 41.3 MB (13.75%) - LeetSync

+++ 3279-alice-and-bob-playing-flower-game/alice-and-bob-playing-flower-game.java
@@ -1,5 +1,5 @@
 class Solution {
     public long flowerGame(int n, int m) {
-        return (long)( (long)( (long)((n+1)/2) * (long)(m/2) )+ (long)( (long)(n/2)* (long)( (m+1)/2 ) ) );
+        return (long)((n + 1) / 2) * (m / 2) + (long)(n / 2) * ((m + 1) / 2);
     }
 }
